Delete my-solution directory
I had my index.html and style.css in here along with the directory that held the screenshots of my solution. Since I've moved the index.html and style.css to the nft-card directory, I'm deleting this folder and creating a new one which holds the screenshots, rather than having this directory which holds another directory with the screenshots.